backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Vector database (FAISS)
faiss-cpu==1.7.4
numpy==1.24.3

# Knowledge Graph (Neo4j)
neo4j==5.15.0

# HTTP client for API calls
httpx==0.25.2

# PDF processing
PyPDF2==3.0.1

# Logging and utilities
python-dotenv==1.0.0

